refactor: extract image-build module from docker (#120)
Move image lifecycle functions (build_base_image, build_derived_image,
derived_image_name) and their supporting helpers (fnv1a_hash, image_label,
image_exists) into a new image_build module.

Introduce BuildConfig as the focused struct carrying only image-build
fields (capsule_dir, pwd, rebuild), replacing individual function parameters.

Expose fnv1a_hash publicly with unit tests for determinism, collision
resistance, known empty-string value, and 16-char hex output format.

Move related tests from docker.rs to image_build.rs; update all callers
in run.rs and tests/docker.rs to use the new module and struct.
