+++ src/docker.rs
@@ -1,179 +1,12 @@
 use crate::stream_parser::StreamParser;
-use anyhow::{bail, Context, Result};
+use anyhow::{Context, Result};
 use std::io::{BufRead, BufReader, Write};
 use std::path::PathBuf;
 use std::process::{Command, Stdio};
 use std::sync::{Arc, Mutex};
 
-/// The base Dockerfile embedded at compile time.
-pub const DOCKERFILE: &str = include_str!("../base-image/Dockerfile");
-
-/// The container entrypoint script embedded at compile time.
-pub const ENTRYPOINT_SH: &str = include_str!("../base-image/entrypoint.sh");
-
 /// The jq stream-display filter embedded at compile time.
 pub const STREAM_DISPLAY_JQ: &str = include_str!("../base-image/stream_display.jq");
-
-const BASE_IMAGE: &str = "capsule";
-const DOCKERFILE_HASH_LABEL: &str = "capsule.dockerfile.hash";
-
-fn dockerfile_hash(content: &str) -> String {
-    // FNV-1a with hardcoded constants — DefaultHasher is not stable across Rust versions.
-    let mut hash: u64 = 0xcbf29ce484222325;
-    for byte in content.bytes() {
-        hash ^= byte as u64;
-        hash = hash.wrapping_mul(0x100000001b3);
-    }
-    format!("{hash:016x}")
-}
-
-fn image_label(name: &str, label: &str) -> Option<String> {
-    let out = Command::new("docker")
-        .args([
-            "image",
-            "inspect",
-            "--format",
-            &format!("{{{{index .Config.Labels \"{label}\"}}}}"),
-            name,
-        ])
-        .output()
-        .ok()?;
-    let s = String::from_utf8(out.stdout).ok()?.trim().to_owned();
-    if s.is_empty() {
-        None
-    } else {
-        Some(s)
-    }
-}
-
-fn image_exists(name: &str) -> bool {
-    Command::new("docker")
-        .args(["image", "inspect", name])
-        .stdout(Stdio::null())
-        .stderr(Stdio::null())
-        .status()
-        .map(|s| s.success())
-        .unwrap_or(false)
-}
-
-/// Returns the derived image name for the given working directory.
-///
-/// Format: `capsule-<basename(pwd)>`. Falls back to `capsule-project` when the
-/// directory has no file-name component (e.g. `/`).
-pub fn derived_image_name(pwd: &std::path::Path) -> String {
-    let basename = pwd
-        .file_name()
-        .and_then(|n| n.to_str())
-        .unwrap_or("project");
-    format!("capsule-{basename}")
-}
-
-/// Build a derived Docker image from `${capsule_dir}/Dockerfile` if it exists.
-///
-/// Returns `Ok(None)` when no `Dockerfile` is found in `capsule_dir`.
-/// Returns `Ok(Some(name))` with the derived image name when the image exists or
-/// was successfully built.
-///
-/// The derived image is named `capsule-<basename(pwd)>` and uses `capsule_dir`
-/// as its build context so relative `COPY` instructions resolve correctly.
-///
-/// If `rebuild` is `false` and the derived image already exists, the build is
-/// skipped and the cached image name is returned.
-pub fn build_derived_image(
-    capsule_dir: &std::path::Path,
-    pwd: &std::path::Path,
-    rebuild: bool,
-) -> Result<Option<String>> {
-    let dockerfile = capsule_dir.join("Dockerfile");
-    if !dockerfile.exists() {
-        return Ok(None);
-    }
-
-    let name = derived_image_name(pwd);
-    let content = std::fs::read_to_string(&dockerfile)
-        .with_context(|| format!("failed to read {}", dockerfile.display()))?;
-    let hash = dockerfile_hash(&content);
-
-    if !rebuild && image_exists(&name) {
-        if image_label(&name, DOCKERFILE_HASH_LABEL).as_deref() == Some(&hash) {
-            return Ok(Some(name));
-        }
-        eprintln!("Derived Dockerfile changed — rebuilding {name}…");
-    } else {
-        eprintln!("Building derived image {name}…");
-    }
-
-    let label = format!("{DOCKERFILE_HASH_LABEL}={hash}");
-    let status = Command::new("docker")
-        .args([
-            "build",
-            "-t",
-            &name,
-            "--label",
-            &label,
-            "-f",
-            &dockerfile.to_string_lossy(),
-            &capsule_dir.to_string_lossy(),
-        ])
-        .status()
-        .context("failed to spawn `docker build` for derived image")?;
-
-    if !status.success() {
-        bail!(
-            "docker build for derived image {name} exited with code {}",
-            status.code().unwrap_or(-1)
-        );
-    }
-
-    eprintln!("Derived image ready.");
-    Ok(Some(name))
-}
-
-/// Build the base `capsule` Docker image from the embedded Dockerfile.
-///
-/// Skips the build when the image exists and its stored hash matches the
-/// embedded Dockerfile. Auto-rebuilds (with layer cache) when the hash
-/// differs. With `rebuild: true`, always rebuilds using `--no-cache`.
-pub fn build_base_image(rebuild: bool) -> Result<()> {
-    let hash = dockerfile_hash(&format!("{DOCKERFILE}{ENTRYPOINT_SH}"));
-
-    if !rebuild && image_exists(BASE_IMAGE) {
-        if image_label(BASE_IMAGE, DOCKERFILE_HASH_LABEL).as_deref() == Some(&hash) {
-            return Ok(());
-        }
-        eprintln!("Base Dockerfile changed — rebuilding {BASE_IMAGE}…");
-    } else {
-        eprintln!("Building {BASE_IMAGE} image…");
-    }
-
-    let ctx = tempfile::tempdir().context("failed to create build context tempdir")?;
-    std::fs::write(ctx.path().join("Dockerfile"), DOCKERFILE)
-        .context("failed to write Dockerfile to build context")?;
-    std::fs::write(ctx.path().join("entrypoint.sh"), ENTRYPOINT_SH)
-        .context("failed to write entrypoint.sh to build context")?;
-
-    let ctx_path = ctx.path().to_string_lossy().into_owned();
-    let label = format!("{DOCKERFILE_HASH_LABEL}={hash}");
-    let mut build_args = vec!["build", "-t", BASE_IMAGE, "--label", &label];
-    if rebuild {
-        build_args.push("--no-cache");
-    }
-    build_args.push(&ctx_path);
-
-    let status = Command::new("docker")
-        .args(&build_args)
-        .status()
-        .context("failed to spawn `docker build`")?;
-    if !status.success() {
-        bail!(
-            "docker build exited with code {}",
-            status.code().unwrap_or(-1)
-        );
-    }
-
-    eprintln!("Image ready.");
-    Ok(())
-}
 
 /// Configuration for a single iteration's `docker run`.
 #[derive(Default, Clone)]
@@ -638,18 +471,6 @@ mod tests {
     // ── Embedded assets ──────────────────────────────────────────────────────
 
     #[test]
-    fn embedded_dockerfile_is_non_empty() {
-        assert!(
-            !DOCKERFILE.is_empty(),
-            "embedded Dockerfile must not be empty"
-        );
-        assert!(
-            DOCKERFILE.contains("FROM archlinux"),
-            "Dockerfile must start from archlinux base"
-        );
-    }
-
-    #[test]
     fn embedded_stream_display_jq_is_non_empty() {
         assert!(
             !STREAM_DISPLAY_JQ.is_empty(),
@@ -688,34 +509,6 @@ mod tests {
             name.ends_with("-3"),
             "name should end with iteration number: {name}"
         );
-    }
-
-    // ── derived_image_name ───────────────────────────────────────────────────
-
-    #[test]
-    fn derived_image_name_uses_basename_of_pwd() {
-        let dir = tempfile::tempdir().expect("temp dir");
-        let project_dir = dir.path().join("my-project");
-        std::fs::create_dir(&project_dir).unwrap();
-        let name = derived_image_name(&project_dir);
-        assert_eq!(name, "capsule-my-project");
-    }
-
-    #[test]
-    fn derived_image_name_handles_root_or_unnamed() {
-        let name = derived_image_name(std::path::Path::new("/"));
-        assert!(name.starts_with("capsule-"), "name={name}");
-    }
-
-    // ── build_derived_image (no Docker) ──────────────────────────────────────
-
-    #[test]
-    fn build_derived_image_returns_none_when_no_dockerfile() {
-        let capsule_dir = tempfile::tempdir().expect("temp dir");
-        let pwd = tempfile::tempdir().expect("temp dir");
-        let result = build_derived_image(capsule_dir.path(), pwd.path(), false)
-            .expect("should not error when Dockerfile absent");
-        assert!(result.is_none(), "expected None when no Dockerfile");
     }
 
     // ── build_docker_args ────────────────────────────────────────────────────
